test(Lotto): lottery & create lotto test
